Move DroppedVariableStats code to lib/IR from lib/Passes.
Also Rewrite DroppedVariableStats to reduce .h file size

The DroppedVariableStats.h file include pulls in a lot of code and leads
to a significant compilation slowdown if included in
MachineFunctionPass.h, which is needed to enable dropped variable
statistics for MIR.

This patch moves all the code to a DroppedVariableStats.cpp file and
replaces those expensive includes with forward declarations to resolve
the issue.
