+++ llvm/include/llvm/IR/DroppedVariableStatsIR.h
@@ -14,11 +14,16 @@
 #ifndef LLVM_CODEGEN_DROPPEDVARIABLESTATSIR_H
 #define LLVM_CODEGEN_DROPPEDVARIABLESTATSIR_H
 
-#include "llvm/IR/InstIterator.h"
-#include "llvm/IR/Module.h"
-#include "llvm/Passes/DroppedVariableStats.h"
+#include "llvm/IR/DroppedVariableStats.h"
 
 namespace llvm {
+
+class Any;
+class StringRef;
+class PassInstrumentationCallbacks;
+class Function;
+class Module;
+class DILocation;
 
 /// A class to collect and print dropped debug information due to LLVM IR
 /// optimization passes. After every LLVM IR pass is run, it will print how many
@@ -28,56 +33,42 @@ public:
   DroppedVariableStatsIR(bool DroppedVarStatsEnabled)
       : llvm::DroppedVariableStats(DroppedVarStatsEnabled) {}
 
-  void runBeforePass(StringRef P, Any IR) {
-    setup();
-    if (const auto *M = unwrapIR<Module>(IR))
-      return this->runOnModule(P, M, true);
-    if (const auto *F = unwrapIR<Function>(IR))
-      return this->runOnFunction(P, F, true);
-  }
+  void runBeforePass(StringRef P, Any IR);
 
-  void runAfterPass(StringRef P, Any IR) {
-    if (const auto *M = unwrapIR<Module>(IR))
-      runAfterPassModule(P, M);
-    else if (const auto *F = unwrapIR<Function>(IR))
-      runAfterPassFunction(P, F);
-    cleanup();
-  }
+  void runAfterPass(StringRef P, Any IR);
 
   void registerCallbacks(PassInstrumentationCallbacks &PIC);
 
 private:
   const Function *Func;
 
-  void runAfterPassFunction(StringRef PassID, const Function *F) {
-    runOnFunction(PassID, F, false);
-    calculateDroppedVarStatsOnFunction(F, PassID, F->getName().str(),
-                                       "Function");
-  }
+  void runAfterPassFunction(StringRef PassID, const Function *F);
 
-  void runAfterPassModule(StringRef PassID, const Module *M) {
-    runOnModule(PassID, M, false);
-    calculateDroppedVarStatsOnModule(M, PassID, M->getName().str(), "Module");
-  }
+  void runAfterPassModule(StringRef PassID, const Module *M);
+
   /// Populate DebugVariablesBefore, DebugVariablesAfter, InlinedAts before or
   /// after a pass has run to facilitate dropped variable calculation for an
   /// llvm::Function.
   void runOnFunction(StringRef PassID, const Function *F, bool Before);
+
   /// Iterate over all Instructions in a Function and report any dropped debug
   /// information.
   void calculateDroppedVarStatsOnFunction(const Function *F, StringRef PassID,
                                           StringRef FuncOrModName,
                                           StringRef PassLevel);
+
   /// Populate DebugVariablesBefore, DebugVariablesAfter, InlinedAts before or
   /// after a pass has run to facilitate dropped variable calculation for an
   /// llvm::Module. Calls runOnFunction on every Function in the Module.
   void runOnModule(StringRef PassID, const Module *M, bool Before);
+
   /// Iterate over all Functions in a Module and report any dropped debug
   /// information. Will call calculateDroppedVarStatsOnFunction on every
   /// Function.
   void calculateDroppedVarStatsOnModule(const Module *M, StringRef PassID,
                                         StringRef FuncOrModName,
                                         StringRef PassLevel);
+
   /// Override base class method to run on an llvm::Function specifically.
   virtual void
   visitEveryInstruction(unsigned &DroppedCount,
@@ -90,10 +81,7 @@ private:
       DenseMap<StringRef, DenseMap<VarID, DILocation *>> &InlinedAtsMap,
       StringRef FuncName, bool Before) override;
 
-  template <typename IRUnitT> static const IRUnitT *unwrapIR(Any IR) {
-    const IRUnitT **IRPtr = llvm::any_cast<const IRUnitT *>(&IR);
-    return IRPtr ? *IRPtr : nullptr;
-  }
+  template <typename IRUnitT> static const IRUnitT *unwrapIR(Any IR);
 };
 
 } // namespace llvm
