+++ llvm/lib/IR/CMakeLists.txt
@@ -26,6 +26,8 @@ add_llvm_component_library(LLVMCore
   DiagnosticInfo.cpp
   DiagnosticPrinter.cpp
   Dominators.cpp
+  DroppedVariableStats.cpp
+  DroppedVariableStatsIR.cpp
   EHPersonalities.cpp
   FPEnv.cpp
   Function.cpp
